refactor(user register): edited controller and writing tests
